Distribution of pos_score
The positive score of tweets seems to be distributed in a highly right-skewed normal distribution manner, with a slight peak occurring around 0.06 - 0.08. However, we can see that a very high number of tweets (around 16000) have a pos_score of 0 indicating that they are either highly negative, negative, or neutral. All the non-positive number of tweets are summed in that one single bar.