Distribution of neg_score
The negative score of tweets seems to be distributed in a highly right-skewed normal distribution manner, with a slight peak occurring around 0.08 - 0.12. However, we can see that a very high number of tweets (around 17500) have a neg_score of 0 indicating that they are either highly positive, positive, or neutral. All the non-negative number of tweets are summed in that one single bar.