Distribution of comp_score for tweets having "Positive" sentiment
For positive tweets, we can see a huge peak in comp_score from around 0.5 and the peak diminishes around 0.75. This could be due to the fact that tweets having a comp_score of more than 0.5 are classified as "Highly Positive" and not just "Positive". However, we could see quite a few tweets which have a comp_score of more than 0.75 being classified as only "Positive" and not "Highly Positive", this could be due to the fact that those tweets might be having a very high neu_score or neg_score.