Distribution of comp_score for tweets having "Extremely Positive" sentiment
In the case of "Highly Positive" tweets, we can see a huge peak jump around 0.75 which diminishes at 1 (because 1 is the max score). This peak takes place right off from the point where the peak of "Positive" tweets end, proving our assumption (comp_score greater than 0.75 being classified as "Highly Positive" tweets) to be correct. We can even spot a few outliers having the comp_score of around -0.75, which could be due to the fact those tweets have high neg_score and neu_score than the pos_score. It could also be the case such tweets are misclassified (possibly due to some error in the initial classifying system)