Distribution of comp_score for tweets having "Neutral" sentiment
We can see that the peak of this graph is wobbly and is on the negative side. There is no "one peak", however, we can say that the region of peak starts from around -0.25 and ends a little bit before -0.75. This could be due to the fact that the statements having a comp_score lower than -0.75 are classified as "Highly Negative" instead of just "Negative"